new packages and rename accounts component to solve a problem
