+++ .meteor/versions
@@ -2,6 +2,7 @@ accounts-base@1.2.15
 accounts-password@1.3.4
 accounts-ui@1.1.9
 accounts-ui-unstyled@1.2.1
+aldeed:browser-tests@0.1.0
 allow-deny@1.0.5
 autoupdate@1.3.12
 babel-compiler@6.18.0-beta.4
@@ -23,6 +24,7 @@ ddp-rate-limiter@1.0.7
 ddp-server@1.3.14-beta.4
 deps@1.0.12
 diff-sequence@1.0.7
+dispatch:mocha@0.3.0
 ecmascript@0.7.1-beta.4
 ecmascript-runtime@0.3.15
 ejson@1.0.13
@@ -74,6 +76,7 @@ npm-mongo@2.2.24
 observe-sequence@1.0.16
 ordered-dict@1.0.9
 peerlibrary:assert@0.2.5
+practicalmeteor:mocha-core@1.0.1
 promise@0.8.8
 random@1.0.10
 rate-limit@1.0.8
